requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
chromadb==0.4.22
neo4j==5.16.0
aiosqlite==0.19.0

# ML & NLP
sentence-transformers==2.3.1
hdbscan==0.8.33
numpy==1.26.3

# Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
